+++ default.nix
@@ -1,19 +1,9 @@
-{ nixpkgs ? import <nixpkgs> {}, compiler ? "default", test ? false }:
+{ nixpkgs ? import <nixpkgs> {}, compiler ? "ghc802" }:
 
-let
-
-  inherit (nixpkgs) pkgs;
-
-  haskellPackages = if compiler == "default" then pkgs.haskellPackages else pkgs.haskell.packages.${compiler};
-
-  drv = with pkgs.haskell.lib; overrideCabal (
-    (haskellPackages.callCabal2nix "stylish-cabal" ./. {})
-      .overrideScope (self: super: { Cabal = null; }))
-    (drv: pkgs.lib.optionalAttrs test {
-      pname = drv.pname + "-${compiler}";
-      configureFlags = [ "-ftest-strictness" "-fwerror" ];
-    });
-
-in
-
-  if pkgs.lib.inNixShell then drv.env else drv
+with nixpkgs.pkgs; stdenv.mkDerivation {
+  name = "shellenv";
+  buildInputs = [
+    zlib
+    haskell.compiler."${compiler}"
+  ];
+}