+++ src/Render.hs
@@ -8,7 +8,6 @@ module Render
     ) where
 
 import Data.List.Compat hiding (group)
-import Data.List.Split
 import Data.Maybe
 import Data.Ord
 import Distribution.Pretty
@@ -60,43 +59,25 @@ fieldValueToDoc k (Field _ f) =
     val' (CabalVersion (Right vr))
         | vr == anyVersion = showVersionRange (orLaterVersion (mkVersion [1,10]))
         | otherwise = showVersionRange vr
-    val' (License l) = pure $ string $ showLicense l
-    val' (SPDXLicense l) = pure $ string $ showLicenseExpr l
+    val' (License l) = pure $ string $ prettyShow l
+    val' (SPDXLicense l) = pure $ string $ prettyShow l
     val' (TestedWith ts) = renderTestedWith ts
     val' (LongList fs) = pure $ vcat $ map filepath fs
     val' (Commas fs) = pure $ fillSep $ punctuate comma $ map filepath fs
     val' (Spaces ls) = pure $ fillSep $ map filepath ls
     val' (Modules ms) = pure $ vcat $ map moduleDoc $ sort ms
     val' (Module m) = pure $ moduleDoc m
-    val' (Extensions es) = val' (LongList $ map showExtension es)
-    val' (FlibType ty) = pure $ string $ showFlibType ty
-    val' (FlibOptions fs) = val' $ Spaces $ map showFlibOpt fs
+    val' (Extensions es) = val' (LongList $ map prettyShow es)
+    val' (FlibType ty) = pure $ string $ prettyShow ty
+    val' (FlibOptions fs) = val' $ Spaces $ map prettyShow fs
     val' x = error $ show x
 fieldValueToDoc k (Description s) = descriptionToDoc k s
 
-parseParagraphs s = map blockToPara $ splitOn "\n\n" s where
-    blockToPara c@('>':_) = Code $ lines c
-    blockToPara x = Words x
-
-descriptionToDoc k s = do
+descriptionToDoc k paras = do
     n <- asks indentSize
-    return $
-        (<>) colon $
-        nest n $
-        case paragraphs of
-            [Words p]
-                -- i still don't know what this does
-             ->
-                group $
-                flatAlt
-                    (linebreak <> fillSep (map text $ words p))
-                    (indent (k + 1) (string p))
-            xs -> line <> vcat (intersperse (green dot) (map paragraph xs))
+    return $ (<>) colon $ nest n $ flatAlt (linebreak <> ds) (indent (k + 1) ds)
   where
-    paragraphs = parseParagraphs s
-    paragraph (Words t) = fillSep (map text $ words t)
-    -- preserve formatting for code blocks
-    paragraph (Code ss) = vcat (map string ss)
+    ds = renderDescription paras
 
 mixinsToDoc k bs
     | k == 0 = pure $ deps ": "