+++ src/Types/Field.hs
@@ -44,6 +44,8 @@ import Distribution.Types.Mixin
 import Distribution.Types.ModuleReexport
 import Distribution.Types.PkgconfigDependency
 import Distribution.Version
+import Documentation.Haddock.Parser
+import Documentation.Haddock.Types (_doc, DocH)
 import Language.Haskell.Extension
 import Prelude.Compat
 
@@ -72,7 +74,7 @@ data FieldVal
     deriving (Show)
 
 data Field
-    = Description String
+    = Description (DocH () Identifier)
     | Field String
             FieldVal
     deriving (Show)
@@ -129,7 +131,7 @@ fieldName (Field s _) = s
 fieldName (Description _) = "description"
 
 desc [] = Nothing
-desc vs = Just (Description vs)
+desc vs = Just (Description $ _doc $ parseParas vs)
 
 buildDeps [] = Nothing
 buildDeps vs = dependencies "build-depends" vs